to long

+++ src/main/java/com/jaworski/serialprotocol/service/WebSocketPublisher.java
@@ -3,5 +3,5 @@ package com.jaworski.serialprotocol.service;
 public interface WebSocketPublisher {
 
     void publishForAllClients(String message);
-    int sessionsCount();
+    long sessionsCount();
 }
